requirements.txt
requests==2.31.0
flask==2.3.3
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.100.0
gunicorn==21.2.0
PyPDF2==3.0.1
python-magic==0.4.26
mega.py==1.0.8